Added test databases, so the tests no longer use the meteorpi.fdb database, instead acquiring their own meteorpi_test1 and meteorpi_test2. Also removed defaults from both MeteorDatabase and MeteorServer, as they were never really the place to have them - you should really know where you're putting your data and not assume a default. Refactored some test code to suit.
Also removed the 'validTo' field in the database for CameraStatus, this is still present in the model but isn't writable, instead being derived from the validFrom date of the subsequence status block. This simplifies a lot of the database logic, and allows for easier insertion of new camera status objects during replication. It means we can no longer have 'holes' in our state, but that's fine.

Factored out CameraStatus db handling into a generator to match the other large domain objects.